core: remove old archive types arch
